Don't init onPersonalized subject to current state of isPersonalized.
Instead init to null and invoke notifyPersonalized method for cases
where we are already personalized.  This ensures that the STOMP
subscribe doesn't occur until after personalization task has run instead
of in parallel. Without this change, the STOMP subscribe could occur
(and was occurring) before personalization had completed and also before
other configuration had taken place, such as overriding of the
compatibilityVersion (which is needed for initial STOMP subscribe).