Set response to null after publishing message.
Fixes issue I noticed in the logs where we were sending form data from a
previous screen as the payload on a subsequent dialog action. Issue was
reproducible when Closing register, then choose 'POS', then login, then
prompted to open register.  The last dialog to open the register was
also including the user login info.  This was a regression from
refactoring dialogs to be handled by the DialogService.