Address issue caused by cordova-monkey-patch-fix where deviceready...
... is no longer received after cordova has sent the event once. Cordova
will normally immediately callback to 'deviceready' subscribers even
after deviceready has already been sent once by cordova. Something about
the monkey patch has broken that behavior.  This ensures that barcode
scan fields on the dynamic-form-field, will still be able to get the
bardcode scan plugin after cordova has initialized.  In a scenario where
a screen with a scan field on it is the first screen shown after
(re)launching the application, it was possible for a race condition to
occur where the barcode scan plugin is attempted to be fetched before it
has been added to the plugin service in the plugin service
initialization.