move screen creation into a directive
This removes the need for the template directive and then locating it with a view child.
I also made the template optional so if you don't have a template it just puts the screen straight on the page.
The classes and themes now get added to the component through the render