No need to use [checked] directive on a reactive form checkbox field
Having [checked] directive in was causing a side affect of the field to
become unchecked when a dialog was popped over top of the form.  This
didn't show until after commit 0d8feb5 when the disabled state of the
control was also set using the FormControl instead of using a directive.

Exact scenario where this was seen was when the checkbox was both
disabled and checked.  When a dialog popped over top of the screen, the
checkbox would uncheck.

SS-10356