Adding a Screen Part decorator amoung some other changes (#15)
Initially this was to fix an issue with the dynamic form control not detecting changes properly and
screen parts not working properly in dialogs.
- This commit adds a dynamic form screen part that takes in a form name and subscribes for messages based on that name.
- Screen parts in dialogs need to subscribe for message type dialog which means that a screen part needs to know if its
host is a screen or a dialog. This is done by allowing PosScreen to inject all of its screen parts as ViewChildren. In
order for all screen parts to be injected by their abstract class they need to provide a reference to themselves by the
ScreenPart token. Since we don't want everyone to have to remember to do this I created a screen part decorator that
will take add the fowardRef provider to the component annotation for you automatically and take in a name to use for
getting the screen part data from the message.
- Once PosScreen has all the screen parts it sets the message type of all its screen parts to its message type.
- Since ViewChildren aren't populated until AfterViewInit we need to make sure that all screen view template
are null protected which is good practice anyways.