Feature/client context (#231)
Whenever an Action request originates from a client device a context object is created and made available to operations resulting from the request.
This is done using a ThreadLocalMap of values. Those values are the sent as headers with and remote service calls and can be injected into endpoints using the @ClientContextProperty attribute.
Also log the request headers on remote service calls so these properties are included in the logging.