Fix ServerLogger to allow logging of objects with circular refs (#403)
- Since we are replacing the console.log and other methods with our own,
this opens us up to how other APIs are using console.log
- Angular is logging some objects via console.log and some of those
objects have circular references. It's possible that other consumers of
our API could also do the same thing.
- When attempting to post a log message that is actually an object with
a circular reference, it causes an error when the HttpClient invokes
JSON.stringify (i.e., 'TypeError: Converting circular structure to
JSON'). Since we send log entries in batches (due to the period between
transmissions), we tend to lose an entire batch of log messages from the
client to the server because of one or more messages in the batch that
may have a circular reference.
- So my solution was, for the case when we intercept a logging method
call that has an object for a 'message', run some code that drops out
the circular references.
- Also added a unit test, server-logger-it.spec.ts, which confirms
proper behavior with a circular reference. I made it an integration test
because it uses both the ConsoleIntercepter and ServerLogger in order to
properly test behavior. The console.log method has to be used to test
logging of an object since the ServerLogger methods do not support
logging of an object at compile time (it only supports strings), but
Javascript allows it at runtime.