Adding support for display of a timer countdown on dialogs (#553)
* Adding support for display of a timer countdown on dialogs.

- After an idle period on the Self Checkout sell screen, a retailer
wanted to automatically display a dialog asking the customer if they
intend to continue or cancel the transaction.
- If the customer does not confirm they want to continue the transaction
within a configurable time period, the retailer wants to cancel the
transaction.
- I thought it would be nice to be able to display a counter in the
dialog with some optional text that shows how much time is remaining
before the transaction would be cancelled.  This has come up before for
other dialogs, but they were not customer facing.
- I contemplated just leveraging the sessionTimeoutMillis for the
dialog, but that is handled on the server side.  The client doesn't have
access to how much time remains before the timeout is triggered.  I also
thought perhaps I would make a subclass of DialogUIMessage to add the
timer functionality, but this resulted in a lot of duplicated template
code on the client side.  Ultimately, I settled on making a message part
of out this.
- An example usage would be as follows:  Add a DialogTimerPart to your
dialog on the server side. Configure it with a timeout value of 30 secs
and a timeoutAction of 'MyTimeout'.  After the dialog is displayed on
the client side, a timer is started and is triggered 30 seconds later.
Upon triggering, the dialog will be automatically closed and the
'MyTimeout' action will be sent back to the server.

* Change over from a MessagePart to a component

- Remove dialog-timer
- Add button-action-timer