Changes To Use Shadow Tables When Configured
Create shadow tables as per configuration in application.yml. Populate the Device Mode selection into the ClientContext and pass it down to DatabaseSchema at initialization. Handle the Device Mode setting smilar to appId and deviceId so that it carries across threads as appropriate. Use shadow table model class in place of the normal table when one is present.