Fix race condition between DialogService and OpenposScreenOutletDirective
- If a dialog is starting to open and a screen is shown immediately
afterwards, there is a possibility in the prior implementation that the
screen will finish showing first, followed by the dialog (which actually
was received before the screen).
- This change uses the dialogOpening flag to ensure that the openDialog
method does not run unless dialogOpening is still true.
- It also updates a couple of places in the closeDialog code to ensure
that dialogOpening is set to false, which will keep a dialog that is in
the process of opening, but has been closed by another caller (e.g.,
OpenposScreenOutletDirective), from actually opening upon execution of
the openDialog method.